requirements.txt
flask
python-dotenv
openai
mysql
mysql-connector-python
bcrypt